Add InfoCURL method

+++ src/evals/metrics/mia/min_k_plus_plus.py
@@ -31,8 +31,9 @@ class MinKPlusPlusAttack(MinKProbAttack):
         # Handle numerical stability
         sigma = torch.clamp(sigma, min=1e-6)
         scores = (
-            target_prob.float().cpu().numpy() - mu.float().cpu().numpy()
-        ) / torch.sqrt(sigma).cpu().numpy()
+            target_prob.to(torch.float32).cpu().numpy()
+            - mu.to(torch.float32).cpu().numpy()
+        ) / torch.sqrt(sigma).to(torch.float32).cpu().numpy()
 
         # Take bottom k% as the attack score
         num_k = max(1, int(len(scores) * self.k))
